moved msg_liquid_stake
